PR changes (Extract Compose Page Objects to common espresso repo, rename package names to be more accurate and structured).
