+++ apps/student/src/androidTest/java/com/instructure/student/ui/utils/StudentComposeTest.kt
@@ -20,9 +20,9 @@ package com.instructure.student.ui.utils
 
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import com.instructure.student.activity.LoginActivity
-import com.instructure.student.ui.pages.compose.CalendarEventDetailsPage
-import com.instructure.student.ui.pages.compose.CalendarEventPage
-import com.instructure.student.ui.pages.compose.CalendarScreenPage
+import com.instructure.canvas.espresso.common.pages.compose.CalendarEventCreateEditPage
+import com.instructure.canvas.espresso.common.pages.compose.CalendarEventDetailsPage
+import com.instructure.canvas.espresso.common.pages.compose.CalendarScreenPage
 import org.junit.Rule
 
 abstract class StudentComposeTest : StudentTest() {
@@ -30,7 +30,7 @@ abstract class StudentComposeTest : StudentTest() {
     @get:Rule(order = 1)
     val composeTestRule = createAndroidComposeRule<LoginActivity>()
 
-    val calendarEventPage = CalendarEventPage(composeTestRule)
+    val calendarEventCreateEditPage = CalendarEventCreateEditPage(composeTestRule)
     val calendarScreenPage = CalendarScreenPage(composeTestRule)
     val calendarEventDetailsPage = CalendarEventDetailsPage(composeTestRule)
 }
