+++ automation/espresso/src/main/kotlin/com/instructure/canvas/espresso/common/interaction/ToDoDetailsInteractionTest.kt
@@ -12,13 +12,13 @@
  *     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  *     See the License for the specific language governing permissions and
  *     limitations under the License.
- */package com.instructure.canvas.espresso.common
+ */package com.instructure.canvas.espresso.common.interaction
 
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import com.instructure.canvas.espresso.CanvasComposeTest
-import com.instructure.canvas.espresso.common.page.ToDoDetailsPage
+import com.instructure.canvas.espresso.common.pages.compose.ToDoDetailsPage
 import com.instructure.canvas.espresso.mockCanvas.MockCanvas
 import com.instructure.canvas.espresso.mockCanvas.addPlannable
 import com.instructure.canvasapi2.models.PlannableType
