+++ automation/espresso/src/main/kotlin/com/instructure/canvas/espresso/common/pages/compose/CalendarEventDetailsPage.kt
@@ -14,7 +14,7 @@
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-package com.instructure.teacher.ui.pages.compose
+package com.instructure.canvas.espresso.common.pages.compose
 
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.assertIsNotDisplayed
